refactor: moved UserInterface.java from InterfaceAdapters package to Frameworks&Drives package update CRC cards in the Phase 2 package
