+++ Phase 2/CRC Card/ApplicationBusinessRules/InstructorFilter.txt
@@ -1,0 +1,10 @@
+Class name: InstructorFilter
+Parent Class: Filter
+Subclass: None
+
+Responsibility:
+- Store the names of profs that the user wishes to avoid as an Arraylist of String
+- Rule out the unwanted timetables delivered by unwanted instructors
+
+Collaborator:
+* Timetable
