+++ Phase 2/CRC Card/EnterpriseBusinessRules/Session.txt
@@ -1,0 +1,8 @@
+Class name: Session
+Parent class: None
+Subclass: None
+
+Responsibility:
+* Store information for one session
+
+Collaborator：
