+++ Phase 2/CRC Card/InterfaceAdapters/MakeCSV.txt
@@ -1,0 +1,10 @@
+Class name: MakeCSV
+Parent class: None
+Subclass: None
+
+Responsibility:
+* Presenting the user with all possible timetables in a CSV format
+
+Collaborator：
+* Timetable
+* Controller
