+++ src/main/java/InterfaceAdapters/Controller.java
@@ -8,6 +8,7 @@ import ApplicationBusinessRule.filter.TimeslotFilter;
 import EnterpriseBusinessRules.NewCourse;
 import FrameworksDrivers.MakeCSV;
 import FrameworksDrivers.UserData;
+import FrameworksDrivers.UserInterface;
 import FrameworksDrivers.WebParse;
 import javafx.application.Application;
 import javafx.collections.FXCollections;
